adding COOKIES[:broadcast_id]

+++ app/controllers/events_controller.rb
@@ -40,26 +40,28 @@ class EventsController < ApplicationController
   def show
     @event = Event.last
 
-    if running_rpi?
-        broadcast_id_fn = `ls /home/pi/rails/tam4dummies/stream/create_broadcast/ | awk '{ print $NF }'`
-    else
-        broadcast_id_fn = `ls /Users/paulonegrao/codecore/railsdir/tam_for_dummies_app/stream/create_broadcast/ | awk '{ print $NF }'`
-    end
-
-    if broadcast_id_fn != ""
-        fn_array = broadcast_id_fn.split("@@t4d@@")
-        stream_id = fn_array[0]
-        broadcast_id = fn_array[1]
-        stream_name = fn_array[2]
-        stream = Stream.find stream_id
-        stream.update_attributes(:broadcast_id => broadcast_id, :stream_name => stream_name)
+    if request.protocol.include?("https")
         if running_rpi?
-          `rm /home/pi/rails/tam4dummies/stream/create_broadcast/*`
+            broadcast_id_fn = `ls /home/pi/rails/tam4dummies/stream/create_broadcast/ | awk '{ print $NF }'`
         else
-          `rm /Users/paulonegrao/codecore/railsdir/tam_for_dummies_app/stream/create_broadcast/*`
+            broadcast_id_fn = `ls /Users/paulonegrao/codecore/railsdir/tam_for_dummies_app/stream/create_broadcast/ | awk '{ print $NF }'`
         end
 
-        session[:broadcast_id] = broadcast_id
+        if broadcast_id_fn != ""
+            fn_array = broadcast_id_fn.split("@@t4d@@")
+            stream_id = fn_array[0]
+            broadcast_id = fn_array[1]
+            stream_name = fn_array[2]
+            stream = Stream.find stream_id
+            stream.update_attributes(:broadcast_id => broadcast_id, :stream_name => stream_name)
+            if running_rpi?
+              `rm /home/pi/rails/tam4dummies/stream/create_broadcast/*`
+            else
+              `rm /Users/paulonegrao/codecore/railsdir/tam_for_dummies_app/stream/create_broadcast/*`
+            end
+
+            cookies[:broadcast_id] = broadcast_id
+        end
     end
 
     respond_to do |format|
